+++ src/main/java/kr/unideal/server/backend/domain/comment/service/CommentService.java
@@ -4,11 +4,11 @@ import jakarta.validation.Valid;
 import kr.unideal.server.backend.domain.comment.controller.dto.request.CommentRequest;
 import kr.unideal.server.backend.domain.comment.controller.dto.request.CommentUpdatedRequest;
 import kr.unideal.server.backend.domain.comment.controller.dto.response.CommentResponse;
-import kr.unideal.server.backend.domain.comment.domain.Comment;
+import kr.unideal.server.backend.domain.comment.entity.Comment;
 import kr.unideal.server.backend.domain.comment.repository.CommentRepository;
-import kr.unideal.server.backend.domain.post.domain.Post;
+import kr.unideal.server.backend.domain.post.entity.Post;
 import kr.unideal.server.backend.domain.post.repository.PostRepository;
-import kr.unideal.server.backend.domain.user.domain.User;
+import kr.unideal.server.backend.domain.user.entity.User;
 import kr.unideal.server.backend.domain.user.repository.UserRepository;
 import kr.unideal.server.backend.global.exception.CustomException;
 import org.springframework.stereotype.Service;
