+++ src/main/java/kr/unideal/server/backend/domain/post/repository/PostRepository.java
@@ -1,7 +1,7 @@
 package kr.unideal.server.backend.domain.post.repository;
 
-import kr.unideal.server.backend.domain.category.domain.Category;
-import kr.unideal.server.backend.domain.post.domain.Post;
+import kr.unideal.server.backend.domain.category.entity.Category;
+import kr.unideal.server.backend.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
